@@ -9,6 +9,7 @@
   "private": true,
   "devDependencies": {
     "@babel/plugin-proposal-optional-chaining": "^7.18.9",
+    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
     "@ethersproject/abi": "5.6.4",
     "@ethersproject/address": "5.6.1",
     "@ethersproject/bignumber": "5.6.2",
@@ -20,7 +21,7 @@
     "@ethersproject/wallet": "5.6.2",
     "@gnosis.pm/safe-apps-web3-react": "0.6.8",
     "@metamask/jazzicon": "2.0.0",
-    "@pangolindex/components": "6.0.1-rc.1",
+    "@pangolindex/components": "6.0.1-rc.3",
     "@pangolindex/exchange-contracts": "2.1.7",
     "@pangolindex/governance": "1.2.0",
     "@pangolindex/sdk": "5.0.0",
@@ -38,7 +39,6 @@
     "@types/react-dom": "16.9.14",
     "@types/react-html-parser": "2.0.2",
     "@types/react-redux": "7.1.23",
-    "@types/react-router-dom": "5.3.3",
     "@types/react-virtualized-auto-sizer": "1.0.1",
     "@types/react-window": "1.8.5",
     "@types/rebass": "4.0.10",
@@ -46,6 +46,7 @@
     "@types/testing-library__cypress": "5.0.9",
     "@typescript-eslint/eslint-plugin": "4.6.0",
     "@typescript-eslint/parser": "4.6.0",
+    "@vitejs/plugin-react": "^3.1.0",
     "@web3-react/abstract-connector": "6.0.7",
     "@web3-react/core": "6.0.9",
     "@web3-react/walletconnect-connector": "6.1.1",
@@ -88,8 +89,7 @@
     "react-markdown": "8.0.1",
     "react-query": "3.34.16",
     "react-redux": "7.2.6",
-    "react-router-dom": "5.3.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "6.8.1",
     "react-spring": "8.0.27",
     "react-use": "15.3.8",
     "react-use-gesture": "6.0.14",
@@ -102,7 +102,9 @@
     "serve": "11.3.2",
     "start-server-and-test": "1.14.0",
     "styled-components": "4.4.1",
-    "typescript": "4.0.5"
+    "typescript": "4.9.5",
+    "vite": "^4.1.4",
+    "vite-plugin-svgr": "^2.4.0"
   },
   "resolutions": {
     "@walletconnect/web3-provider": "1.1.1-alpha.0",
@@ -110,10 +112,8 @@
     "react-error-overlay": "6.0.9"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts --max_old_space_size=8192 build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "tsc && vite build",
     "serve": "serve build -l 3000",
     "version": "npm run build",
     "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --fix",
@@ -147,12 +147,11 @@
     "@babel/core": "^7.20.5",
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
     "@coinbase/cbpay-js": "^1.6.0",
-    "@sentry/react": "6.19.1",
-    "@sentry/tracing": "6.19.1",
+    "@sentry/react": "7.38.0",
+    "@sentry/tracing": "7.38.0",
     "@types/react-portal": "4.0.4",
     "graphql-request": "4.2.0",
     "lightweight-charts": "3.8.0",
-    "near-api-js": "0.44.2",
     "npm-force-resolutions": "0.0.10",
     "patch-package": "6.4.7",
     "react-portal": "4.2.1"
