@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react'
-import { Redirect, Route, Switch } from 'react-router-dom'
+import { Route, Routes } from 'react-router-dom'
 import styled from 'styled-components'
 import { MENU_LINK } from 'src/constants'
 import { useChainId } from 'src/hooks'
@@ -7,10 +7,8 @@ import { shouldHideMenuItem } from 'src/utils'
 const Polling = React.lazy(() => import('../components/Header/Polling'))
 const Popups = React.lazy(() => import('../components/Popups'))
 const Web3ReactManager = React.lazy(() => import('../components/Web3ReactManager'))
-const DarkModeQueryParamReader = React.lazy(() => import('../theme/DarkModeQueryParamReader'))
 const Dashboard = React.lazy(() => import('./Dashboard'))
 const MigrateV2 = React.lazy(() => import('./Migrate'))
-const CustomRoute = React.lazy(() => import('./Route'))
 const Layout = React.lazy(() => import('../layout'))
 const SwapV2 = React.lazy(() => import('./Beta/Swap'))
 const StakeV2 = React.lazy(() => import('./Beta/Stake'))
@@ -54,9 +52,6 @@ interface IRoute {
   menuLink?: MENU_LINK //Bring it blank if it is same with path
   path: string | MENU_LINK
   component: React.FC
-  layout: React.FC
-  strict?: boolean
-  exact?: boolean
 }
 
 export default function App() {
@@ -64,107 +59,77 @@ export default function App() {
   const routes: IRoute[] = [
     {
       path: MENU_LINK.dashboard,
-      component: Dashboard,
-      layout: Layout
+      component: Dashboard
     },
     {
       menuLink: MENU_LINK.migrate,
       path: `${MENU_LINK.migrate}/:version`,
-      component: MigrateV2,
-      layout: Layout
+      component: MigrateV2
     },
     {
       path: MENU_LINK.swap,
-      component: SwapV2,
-      layout: Layout
+      component: SwapV2
     },
     {
       menuLink: MENU_LINK.stake,
       path: `${MENU_LINK.stake}/:version`,
-      strict: true,
-      component: StakeV2,
-      layout: Layout
+      component: StakeV2
     },
     {
       path: MENU_LINK.vote,
-      strict: true,
-      component: GovernanceV2,
-      layout: Layout
+      component: GovernanceV2
     },
     {
       menuLink: MENU_LINK.vote,
       path: `${MENU_LINK.vote}/:id`,
-      strict: true,
-      component: GovernanceDetailV2,
-      layout: Layout
+      component: GovernanceDetailV2
     },
     {
       path: MENU_LINK.pool,
-      component: PoolV2,
-      layout: Layout
+      component: PoolV2
     },
     {
       path: `${MENU_LINK.buy}/:type`,
-      strict: true,
-      component: BuyV2,
-      layout: Layout
+      component: BuyV2
     },
     {
       path: MENU_LINK.bridge,
-      component: BridgeV2,
-      layout: Layout
+      component: BridgeV2
     },
     {
       path: MENU_LINK.airdrop,
-      component: AirdropV2,
-      layout: Layout
+      component: AirdropV2
     },
     {
       path: MENU_LINK.stakev2,
-      component: SarStake,
-      layout: Layout
+      component: SarStake
     }
   ]
 
   return (
     <Suspense fallback={null}>
-      <Route component={DarkModeQueryParamReader} />
       <AppWrapper>
         <BodyWrapper>
           <Popups />
           <Polling />
           <Web3ReactManager>
-            <Switch>
-              {routes.map(
-                (route, i) =>
-                  !shouldHideMenuItem(chainId, route?.menuLink || (route.path as MENU_LINK)) && (
-                    <CustomRoute
-                      key={i}
-                      exact
-                      strict={route?.strict}
-                      path={route.path}
-                      component={route.component}
-                      layout={route.layout}
-                    />
-                  )
-              )}
-              {/* <CustomRoute
-                exact
-                strict
-                path="/beta/stake/:version/:rewardCurrencyId"
-                component={ManageStakeV2}
-                layout={Layout}
-              /> */}
+            <Routes>
+              <Route path="/" element={<Layout />}>
+                {/* render all dynamic routes */}
+                {routes
+                  .filter(route => !shouldHideMenuItem(chainId, route?.menuLink || (route.path as MENU_LINK)))
+                  .map((route, i) => {
+                    const Component = route.component
+                    return <Route key={i} path={route.path} element={<Component />} />
+                  })}
 
-              {/* <Route exact path="/beta/migrate/:version" component={MigrateV2} /> */}
-
-              <CustomRoute exact path="/policy/privacy" component={() => <Policy policy="privacy" />} layout={Layout} />
-              <CustomRoute exact path="/policy/cookie" component={() => <Policy policy="cookie" />} layout={Layout} />
-              <CustomRoute exact path="/policy/terms" component={() => <Policy policy="terms" />} layout={Layout} />
-
-              {/* <Route component={RedirectPathToSwapOnly} /> */}
-              <Redirect to={MENU_LINK.dashboard} />
-            </Switch>
+                <Route index element={<Dashboard />} />
+                <Route path="policy/privacy" element={<Policy policy="privacy" />} />
+                <Route path="policy/cookie" element={<Policy policy="cookie" />} />
+                <Route path="policy/terms" element={<Policy policy="terms" />} />
+                <Route path="*" element={<Dashboard />} />
+              </Route>
+            </Routes>
           </Web3ReactManager>
         </BodyWrapper>
       </AppWrapper>
