@@ -27,7 +27,7 @@ try {
     release: `pangolin-interface@${Package.version}`, //manual for now
     tracesSampleRate: 0.4,
     allowUrls: ['https://app.pangolin.exchange', 'https://dev.pangolin.exchange'],
-    enabled: process.env.NODE_ENV === 'production',
+    enabled: import.meta.env.PROD,
     ignoreErrors: [
       'ResizeObserver loop limit exceeded',
       'Blocked a frame with origin "https://app.pangolin.exchange" from accessing a cross-origin frame.'
@@ -42,7 +42,7 @@ if ('ethereum' in window) {
   ;(window.ethereum as any).autoRefreshOnNetworkChange = false
 }
 
-const mixpanelToken = process.env.REACT_APP_MIXPANEL
+const mixpanelToken = import.meta.env.VITE_MIXPANEL
 
 const queryClient = new QueryClient({
   defaultOptions: {
