@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+import { Outlet } from 'react-router-dom'
 import Header from './Header'
 import Sidebar from './Sidebar'
 import { Wrapper, MainContent, AppContent } from './styled'
@@ -6,7 +7,7 @@ import Footer from './Footer'
 import URLWarning from 'src/components/Header/URLWarning'
 import MobileMenu from './Header/MobileMenu'
 
-const Layout: React.FC<unknown> = ({ children }) => {
+const Layout: React.FC<unknown> = () => {
   const [isDrawerCollapsed, setIsDrawerCollapsed] = useState(true)
 
   const [activeMobileMenu, setActiveMobileMenu] = useState(false)
@@ -25,7 +26,9 @@ const Layout: React.FC<unknown> = ({ children }) => {
       <MainContent collapsed={isDrawerCollapsed}>
         <URLWarning />
         <Header activeMobileMenu={activeMobileMenu} handleMobileMenu={handleMobileMenu} />
-        <AppContent>{children}</AppContent>
+        <AppContent>
+          <Outlet />
+        </AppContent>
         {activeMobileMenu && <MobileMenu activeMobileMenu={activeMobileMenu} handleMobileMenu={handleMobileMenu} />}
         <Footer />
       </MainContent>
