@@ -5,9 +5,7 @@ import CoinbasePay from './CoinbasePay'
 import Moonpay from './Moonpay'
 import C14 from './C14'
 
-export interface BuyProps {
-  type: string
-}
+export type BuyProps = Record<'type', BUY_MENU_LINK>
 
 export default function BuyV2() {
   const params = useParams<BuyProps>()
