@@ -4,7 +4,6 @@ import { NewsWidget, WatchList, Portfolio, useTranslation } from '@pangolindex/c
 import { CHAINS } from '@pangolindex/sdk'
 import { useChainId } from 'src/hooks'
 import { Hidden, Visible } from 'src/theme'
-import { MENU_LINK } from 'src/constants'
 import { WATCHLIST_ACCESS } from 'src/constants/accessPermissions'
 
 const Dashboard = () => {
@@ -23,9 +22,7 @@ const Dashboard = () => {
           <StatsWrapper isShowWatchList={WATCHLIST_ACCESS[chainId]}>
             <Portfolio />
 
-            {WATCHLIST_ACCESS[chainId] && (
-              <WatchList visibleTradeButton={true} tradeLinkUrl={MENU_LINK.swap} redirect={true} />
-            )}
+            {WATCHLIST_ACCESS[chainId] && <WatchList />}
           </StatsWrapper>
         )}
 
